+++ core/components/romanesco/docs/changelog.md
@@ -9,13 +9,14 @@ New features:
 
 Fixes and improvements:
 - Don't remove rows in mobile reversed grids (breaks multi-row nested layouts)
+- Also hide designated mobile elements if parent grid is stackable on tablet
 - Use placeholder to detect ToC instead of toc_enabled_templates system setting
 - Show ToC submenu in dropdown on tablet and mobile
 - Move shared overview settings to separate molecules
 - Fix regClient scripts not being parsed in Markdown pages on first load
 - Load external JS from head with defer
 - Load all conditional assets through a single snippet
-- Don't use tag label variant anymore (excluded from semantic.css)
+- Don't use tag label variant anywhere (excluded from default semantic.css)
 - Add SUI list class to ul and ol tags in Markdown (only at first level)
 - Remove conditionals for empty logo paths (to clear media source path)
 - Add theme variables for inverted logo paths